refactor: 중복으로 사용되고 있던 stepStore을 src root 경로로 분리합니다.
