--- src/apps/survey/stores/step/step.types.ts
@@ -1,8 +1,0 @@
-export interface StepInitialState {
-  currentStep: number;
-}
-
-export interface StepState extends StepInitialState {
-  goToNext: VoidFunction;
-  dispatchInitialize: VoidFunction;
-}
